content/canvas/test/webgl/test_webgl_conformance_test_suite.html

<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN"
  "http://www.w3.org/TR/html4/loose.dtd">
<html>
<head>
<title>
Mochitest version of the WebGL Conformance Test Suite
</title>
<script type="application/javascript" src="/MochiKit/packed.js"></script>
<script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
<script type="text/javascript" src="resources/webgl-test-harness.js"></script>
<script>

function occurencesCount(array, obj) {
  var i = array.length;
  var result = 0;
  while (i--) {
    if (array[i] === obj) {
      ++result;
    }
  }
  return result;
}

/**
 * This is copied from webgl-test-harness.js where it is defined as a private function, not accessible to us (argh!)
 *
 * Loads text from an external file. This function is synchronous.
 * @param {string} url The url of the external file.
 * @return {string} the loaded text if the request is synchronous.
 */
var loadTextFileSynchronous = function(url) {
  var error = 'loadTextFileSynchronous failed to load url "' + url + '"';
  var request;
  if (window.XMLHttpRequest) {
    request = new XMLHttpRequest();
    if (request.overrideMimeType) {
      request.overrideMimeType('text/plain');
    }
  } else {
    throw 'XMLHttpRequest is disabled';
  }
  request.open('GET', url, false);
  request.send(null);
  if (request.readyState != 4) {
    throw error;
  }
  return request.responseText;
};

function start() {

  function webglBranch() {
    return Components.classes["@mozilla.org/preferences-service;1"]
                     .getService(Components.interfaces.nsIPrefService)
                     .getBranch("webgl.");
  }

  function setEnabledForAllSites(value) {
    netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
    webglBranch().setBoolPref("enabled_for_all_sites", value);
  }

  function getEnabledForAllSites() {
    netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
    return webglBranch().getBoolPref("enabled_for_all_sites");
  }

  function setForceOSMesa(value) {
    netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
    webglBranch().setBoolPref("force_osmesa", value);
  }

  function getForceOSMesa() {
    netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
    return webglBranch().getBoolPref("force_osmesa");
  }

  function getMochitestNativeGL() {
    netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
    return webglBranch().getBoolPref("mochitest_native_gl");
  }

  function setOSMesaLib(value) {
    netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
    webglBranch().setCharPref("osmesalib", value);
  }

  function getOSMesaLib() {
    netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
    return webglBranch().getCharPref("osmesalib");
  }

  function restoreOldPrefs() {
    setEnabledForAllSites(saved_enabled_for_all_sites);
    setForceOSMesa(saved_force_osmesa);
    setOSMesaLib(saved_osmesalib);
  }

  function forceOSMesaInThisTestRun() {
    return timesTestSuiteAlreadyRun == 1 ? true : false;
  }

  function GLImplementationName() {
    return forceOSMesaInThisTestRun() ? 'OSMesa' : 'default GL';
  }

  var reportType = WebGLTestHarnessModule.TestHarness.reportType;

  var Page = function(reporter, url) {
    this.reporter = reporter;
    this.url = url;
    this.totalTests = 0;
    this.totalSuccessful = 0;
    this.totalTimeouts = 0;

    var li = reporter.localDoc.createElement('li');
    var div = reporter.localDoc.createElement('div');
    var a = reporter.localDoc.createElement('a');
    a.href = url;
    var node = reporter.localDoc.createTextNode(url);
    a.appendChild(node);
    div.appendChild(a);
    li.setAttribute('class', 'testpage');
    li.appendChild(div);
    var ul = reporter.localDoc.createElement('ul');
    var node = reporter.localDoc.createTextNode('');
    li.appendChild(ul);
    div.appendChild(node);
    this.totalsElem = node;
    this.resultElem = ul;
    this.elem = li;
  };

  Page.prototype.isExpectedToFullyPass = function() {
    return occurencesCount(testsExpectedToFail, this.url) == 0;
  }

  Page.prototype.errormsg = function(msg) {
    return msg + ' (URL: ' + this.url + ', using ' + GLImplementationName() + ')';
  }

  Page.prototype.addResult = function(msg, success) {
    ++this.totalTests;
    if (success === undefined) {
      ++this.totalTimeouts;
      var result = "timeout";
      var css = "timeout";
      // only few timeouts are actually caught here --- most are caught in finishPage().
      if (this.isExpectedToFullyPass()) {
        ok(false, this.errormsg('Test timed out, "' + msg + '"'));
      }
    } else if (success) {
      ++this.totalSuccessful;
      var result = "success";
      var css = "success";
      // don't report success.
      return;
    } else {
      var result = "failed";
      var css = "fail";
      if (this.isExpectedToFullyPass()) {
        ok(false, this.errormsg('Test failed, "' + msg + '"'));
      }
    }

    var node = this.reporter.localDoc.createTextNode(result + ': ' + msg);
    var li = this.reporter.localDoc.createElement('li');
    li.appendChild(node);
    li.setAttribute('class', css);
    this.resultElem.appendChild(li);
  };

  Page.prototype.startPage = function() {
    this.totalTests = 0;
    this.totalSuccessful = 0;
    this.totalTimeouts = 0;
    // remove previous results.
    while (this.resultElem.hasChildNodes()) {
      this.resultElem.removeChild(this.resultElem.childNodes[0]);
    }
    this.totalsElem.textContent = '';
    return true;
  };

  Page.prototype.finishPage = function(success) {
    var msg = ' (' + this.totalSuccessful + ' of ' +
              this.totalTests + ' passed)';
    if (success === undefined) {
      var css = 'testpagetimeout';
      msg = '(*timeout*)';
      ++this.totalTests;
      ++this.totalTimeouts;
      if (this.isExpectedToFullyPass()) {
        ok(false, this.errormsg('Unexpected timeout in this test page'));
        window.dump('WebGL test error: test page timeout: ' + this.url +
                    ' (using ' + GLImplementationName() + ')\n');
      }
    } else if (this.totalSuccessful != this.totalTests) {
      var css = 'testpagefail';
      if (this.isExpectedToFullyPass()) {
        window.dump('WebGL test error: test page failure: ' + this.url +
                    ' (using ' + GLImplementationName() + ')\n');
      }
      // failures have already been reported for the sub-tests
    } else {
      var css = 'testpagesuccess';
      if (forceOSMesaInThisTestRun()) {
        testsSuccessesWithOSMesa.push(this.url);
      }
      if (this.isExpectedToFullyPass()) {
        ok(true, this.errormsg('Successful test page'));
      }
      window.dump('WebGL test page successful: ' + this.url +
                    ' (using ' + GLImplementationName() + ')\n');
    }
    this.elem.setAttribute('class', css);
    this.totalsElem.textContent = msg;
  };

  var Reporter = function() {
    this.localDoc = document;

    this.fullResultsElem
      = document.getElementById(forceOSMesaInThisTestRun() ? "results-osmesa" : "results-default");

    this.resultElem = document.getElementById("results");
    var node = this.localDoc.createTextNode('');
    this.fullResultsElem.appendChild(node);
    this.fullResultsNode = node;
    this.iframe = document.getElementById("testframe");
    this.currentPageElem = null;
    this.totalPages = 0;
    this.pagesByURL = {};
    this.currentPage = null;
    this.totalTests = 0;
    this.totalSuccessful = 0;
    this.totalTimeouts = 0;
  };

  Reporter.prototype.runTest = function(url) {
    var page = this.pagesByURL[url];
    page.startPage();
    this.currentPage = page;
    this.iframe.src = url;
    return result;
  };

  Reporter.prototype.addPage = function(url) {
    this.currentPage = new Page(this, url, this.resultElem);
    this.resultElem.appendChild(this.currentPage.elem);
    ++this.totalPages;
    this.pagesByURL[url] = this.currentPage;
  };

  Reporter.prototype.startPage = function(url) {
    var page = this.pagesByURL[url];
    this.currentPage = page;
    statusTextNode.textContent = 'Running (' + GLImplementationName() + '), URL: ' + url;
    expectedtofailTextNode.textContent = testsExpectedToFail.length +
                                         ' test pages are expected to fail out of ' +
                                         this.totalPages;
    return page.startPage();
  };

  Reporter.prototype.totalFailed = function() {
    return this.totalTests - this.totalSuccessful;
  };

  Reporter.prototype.displayStats = function() {
    this.fullResultsNode.textContent =
      this.totalSuccessful + ' passed, ' +
      this.totalFailed() + ' failed, ' +
      this.totalTimeouts + ' timed out';
  };

  Reporter.prototype.addResult = function(msg, success) {
    if (this.currentPage != null) {
      this.currentPage.addResult(msg, success);
    }
  };

  Reporter.prototype.finishPage = function(success) {
    if (this.currentPage != null) {
      this.currentPage.finishPage(success); // must call that first, since this is where totalTimeouts is computed
      this.totalTests += this.currentPage.totalTests;
      this.totalSuccessful += this.currentPage.totalSuccessful;
      this.totalTimeouts += this.currentPage.totalTimeouts;
      this.currentPage = null;
      this.displayStats();
    }
  };

  Reporter.prototype.completedCurrentTestSuiteRun = function() {
        timesTestSuiteAlreadyRun++;
        runTestSuite();
  };

  Reporter.prototype.reportFunc = function(type, msg, success) {
    switch (type) {
      case reportType.ADD_PAGE:
        return this.addPage(msg);
      case reportType.START_PAGE:
        return this.startPage(msg);
      case reportType.TEST_RESULT:
        return this.addResult(msg, success);
      case reportType.FINISH_PAGE:
        return this.finishPage(success);
      case reportType.FINISHED_ALL_TESTS:
        this.completedCurrentTestSuiteRun();
        return true;
      default:
        throw 'unhandled';
        break;
    };
  };

  function doRunTests() {
    var reporter = new Reporter();

    if (!mochitestNativeGL && !forceOSMesaInThisTestRun()) {
        var errmsg = "Disabled. By default, we don't run tests on native OpenGL as it can give inconsistent results. " +
                  "If you want to run tests on native OpenGL, set the preference webgl.mochitest_native_gl to true.";
        reporter.fullResultsNode.textContent = errmsg;
        reporter.completedCurrentTestSuiteRun();
        return;
    }

    // set WebGL-related preferences needed to run WebGL tests
    setEnabledForAllSites(true);
    setForceOSMesa(forceOSMesaInThisTestRun());

    // try to create a dummy WebGL context, just to catch context creation failures once here,
    // rather than having them result in 100's of failures (one in each test page)
    var canvas = document.getElementById(forceOSMesaInThisTestRun() ? "webglcheck-osmesa" : "webglcheck-default");
    var ctx = null;
    try {
        ctx = canvas.getContext("experimental-webgl");
    } catch(e) {}

    if (ctx) {
        var iframe = document.getElementById("testframe");
        var testHarness = new WebGLTestHarnessModule.TestHarness(
            iframe,
            '00_testFIXME_list.txt',
            function(type, msg, success) {
                return reporter.reportFunc(type, msg, success);
            });
        testHarness.setTimeoutDelay(10000); // and make it much higher when running under valgrind.
        window.webglTestHarness = testHarness;
        testHarness.runTests();
    } else {
        // if the webgl context is null, we agreed that we wouldn't fail the test, we only add a TODO mentioning
        // that the GL setup needs to be fixed in order for these tests to be actually run. The reason why we don't
        // currently want to fail the test is that we want to be able to land this mochitest as soon as possible and
        // just take care of the GL setups on the test machines afterwards.
        var errmsg = "";
        if (forceOSMesaInThisTestRun()) {
            errmsg
                = "Disabled, because a usable OSMesa was not found. " +
                  "Install the OSMesa library and let webgl.osmesalib point to it. " +
                  "This makes the WebGL tests run in a setup that is completely independent from your graphics hardware.";
        } else {
            errmsg
                = "Disabled, because we were unable to create a WebGL context anyhow.";
        }
        reporter.fullResultsNode.textContent = errmsg;
        todo(false, errmsg);
        reporter.completedCurrentTestSuiteRun();
    }
  };

  function runTestSuite() {
    if (timesTestSuiteAlreadyRun == 2) {
      for(var i = 0; i < testsExpectedToFail.length; ++i) {
        if(occurencesCount(testsSuccessesWithOSMesa, testsExpectedToFail[i]) == 1) {
          ok(false, 'The following test page is successful with OSMesa, ' +
                    'remove it from the list of failing pages: ' +
                    testsExpectedToFail[i]);
          window.dump('WebGL test error: page must be removed from failing_tests.txt: ' +
                      testsExpectedToFail[i] + '\n');
        }
      }
      restoreOldPrefs();
      statusTextNode.textContent = 'Finished';
      SimpleTest.finish();
    } else {
      doRunTests();
    }
  };

  var timesTestSuiteAlreadyRun = 0;

  var saved_enabled_for_all_sites = getEnabledForAllSites();
  var saved_force_osmesa = getForceOSMesa();
  var saved_osmesalib = getOSMesaLib();
  var mochitestNativeGL = getMochitestNativeGL();

  SimpleTest.waitForExplicitFinish();
  SimpleTest.requestLongerTimeout(2);

  var statusElem = document.getElementById("status");
  var statusTextNode = document.createTextNode('');
  statusElem.appendChild(statusTextNode);

  var expectedtofailElem = document.getElementById("expectedtofail");
  var expectedtofailTextNode = document.createTextNode('');
  expectedtofailElem.appendChild(expectedtofailTextNode);

  var testsExpectedToFail = loadTextFileSynchronous('failing_tests.txt').split('\n');
  var testsSuccessesWithOSMesa = [];

  runTestSuite();
}

</script>
</head>
<body onload="start()">
<p id="display"></p>
<div id="content" style="display: none">

</div>
<table border="2">
<tr style="height: 300px;">
<td>
<table>
<tr><td><h4>WebGL Conformance Test Runner</h4>
<tr><td>
  <div style="border: 1px">
    <b>Status:</b> <div><span id="status"></span></div><br />
    <b>Results:</b>
    <table>
    <tr>
        <td>With default GL:</td>
        <td><div><span id="results-default"></span></div></td>
    </tr>
    <tr>
        <td>With OSMesa:</td>
        <td><div><span id="results-osmesa"></span></div></td>
    </tr>
    </table>
    <br />
    <div><span id="expectedtofail"></span></div>
  </div>
</td></tr>
</table>
</td>
<td>
<iframe id="testframe" scrolling="yes" width="100%" height="100%"></iframe>
</td>
</tr>
<tr>
<td colspan="2">
<div style="text-align: left; width: 100%; height: 100%; overflow: auto;">
<div><ul id="results"></ul></div>
</div>
</td>
</tr>
</table>
<canvas id="webglcheck-default" style="display: none;"></canvas>
<canvas id="webglcheck-osmesa" style="display: none;"></canvas>
</body>
</html>
